Refactor load of properties
